Moving concurrency value to avoid make problems
When we tell the jdk test makefile to use both CONCURRENCY and EXTRA_JTREG_OPTIONS, it deducts 1 from the concurrency number we're trying to pass in, and uses two identical concurrency options. The latter causes the test to fail, and the former complicates our efforts to control the jtreg run.

(I believe the latter problem may be due to https://bugs.openjdk.java.net/browse/JDK-8157605 )

To solve both problems, I'm changing our shell script so we don't set the CONCURRENCY environment variable at all, and instead we set it directly on the command line via the EXTRA_JTREG_OPTIONS environment variable.