Fetch freemarker.jar if necessary
The openj9 variant currently requires freemarker.jar to be downloaded
before the build starts, so this commit will fetch the appropriate
version of freemarker and makes sure the configure command knows
where it is. Currently uses ALTERNATE_VARIANT, but should be changed
to test JVM_VARIANT once openj9 becomes an OpenJDK JVM variant.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>