Repair JVM_VARIANT bug
The old form of build.sh caused it to ignore the command line when
setting jvm_variant. This has been fixed, so a bug in makejdk.sh
is now causing build failures (makejdk.sh used :server, which is
wrong). This is the fix for that, so makejdk.sh now passes the correct
jvm_variant value.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>