Switch to API v3 and provide better diagnostics
Currently there is logic where when there is no directory
present pointed to by JDKxx_BOOT_DIR a fallback is in
place to download the JDK N-1 boot JDK version via the
API. If this fails for some reason, not much diagnostics
are available.

This patch tries to simplify the logic a bit and adds an
additional 'java -version' output for the - potentially -
changed boot JDK the build will be using.

Signed-off-by: Severin Gehwolf <sgehwolf@redhat.com>