Replace string includes with count
String.includes(String) does not appear to exist, which I'm sure was
a typo, so I'm replacing it with count(String).

String.contains(String) is another option, but I didn't use that
because the API says it isn't mandatory for JDK 1.5+. There's also
String.indexOf, according to the error message, but I'm not seeing
that in the API doc, so let's play it safe.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>