Disable jdk14 build+test, and enable jdk15 build+test
This is being done because we've passed the last release for jdk14,
and we need start testing 15 in preparation for the September release.

Most JDK15 builds are already active, but there are a few platforms
(windows 32bit and XL variants) that needed to be added, based on the
fact that jdk14 was creating builds for them.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>