Remove Node Existence Check Prior to Docker Build
Docker nodes do not exist prior to us trying to use them, as they
are dynamically generated.

So waiting for one to become active prior to use is a futile effort,
and the inevitable timeout causes build failure.

So this change removes the "wait for active node" step from the
docker builds.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>