Give build pipeline priority control over config options
The user should always have the ability to specify options on the
commandline. It's useful for debugging, releasing, prototyping fixes,
etc.

However, giving the user a lower priority than the options provided
by the platform-specific scripts means that the user's options can
be overridden, and the user then has to go digging around in
various files to figure out why the scripts aren't doing why they're
told.

I think this is effort wasted, and the user's commands should always
have priority, even if it breaks things.

This command ensures that the user (or at least the surface params
that the user can trivially see and change in jenkins) always has
priority.

Signed-off-by: Adam Farley <adfarley@redhat.com>