Build cacerts during build from Mozilla's list (#2113)
The cacerts trust store provided by OpenJDK lacks a few common certificates
(see https://github.com/AdoptOpenJDK/openjdk-support/issues/13). We reached
out to OpenJDK, but there was little interest to change the state of affairs
(see https://mail.openjdk.java.net/pipermail/jdk-dev/2020-May/004305.html).
Consequently, the AdoptOpenJDK TSC decided to replace the bundled trust
store with the root CA certificates included in Mozilla Firefox. Mozilla
runs a trusted root CA program and is used by, amongst others, most Linux
distributions. Running our own root CA program was not an option (too much
work, difficult). But we leave the option open to include CA certificates
on a case-by-case basis. If anyone wants to build with the stock OpenJDK
certificates, pass `--custom-cacerts false`.

The list of CA certificates is not downloaded on demand but stored in the
repository. This prevents intermittent download failures and makes it easier
to inspect what certificates we have bundled with a certain release. Changes
are also clearly visible in the revision history. This is also the reason that
the cacerts file is built from source during the JDK build. The downside is
that the list of certificates needs to be regularly updated.

The cacerts trust store is built with keytool from the boot JDK to ensure
that it is compatible with the built JDK version.