Shenandoah present on win32
As per the Shenandoah project webpage, if shenandoah is set as a
with-jvm-features option during pre-build configuration (which Adopt
does), then we should expect Shenandoah to be both present and
functional in the build.

My assumption is that Windows 32bit was simply skipped during smoke
test pre-deployment testing, and this problem was simply not seen
as a result.

Signed-off-by: Adam Farley <adfarley@redhat.com>