Fix smoke test to expect the right vendor name, plus other fixes
The vendor name in the -version output check has been modified to
match the vendor name we're expecting to find after a recent
change to the build configurations.

We also need the hard-coded vendor name to match the vendor name
present in the adoptium implementation of the VmPropertiesChecks
interface (AdoptiumPropertiesChecks).

Lastly, the test has been modified so it compiles.

Signed-off-by: Adam Farley <adfarley@redhat.com>