Enabling Hotspot Smoke Tests and correcting an expected value
This change includes a playlist tweak to enable the hotspot smoke
tests to run, and also modifies the expected system property
"java.vm.vendor" for adoptium's Hotspot JDK to be "Eclipse Foundation".

As this change brings "java.vm.vendor" in line with the values
expected from other properties, I'm asserting that this is likely
not a bug, but I'll consult with my fellow developers before
merging, just to be safe.

Signed-off-by: Adam Farley <adfarley@redhat.com>