Remove check that moves minor to tags in variant JSON (#2832)
This precaution if has been there since it was authored.
I can't understand why it is there. The code seems to work
without it. It seems to unintentionally be moving those
values for every non-milestone release. That is, the code
works for milestones only but for releases it moves minor
to tags.

eg. 0.29.0-m1 becomes 0, 29, 0, m1 which is correct
but 0.29.0 becomes 0, 0, 0, 29 which is incorrect

Related #2077

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>