Temporarily patch architecture confusion when cross-compiling on mac
Currently, when we build an x64 mac build on an arm64 machine, we put
arm64 all over the place by mistake.

This change is to partially compensate for that mistake until such
time as we can impliment a comprehensive seperation of build
architecture and target architecture in our build scripts.

Signed-off-by: Adam Farley <adfarley@redhat.com>