Adding secure mode option to build scripting
This mode will eventually be used to disable a range of
build script functionality that raises the potential security
risk level during the build process.

After this commit is merged, we will need to add the flag to the
pipeline job configurations in order to pass it into build jobs.

Signed-off-by: Adam Farley <adfarley@redhat.com>