Adding github api check to SBOM sha validation (#4338)
* Adding github api check to SBOM sha validation

When checking that the sha provided in the SBOM is valid,
we sometimes encounter an intermittent problem where github
returns an unusual HTML response code even when the commit
exists. The api seemed unaffected.

Here is a backup check that uses the api if the usual url
causes problems again.

Signed-off-by: Adam Farley <adfarley@ibm.com>